move setting to NettyHttpServerConfiguration
