+++ http-server-netty/src/main/java/io/micronaut/http/server/netty/jackson/JsonHttpContentSubscriberFactory.java
@@ -18,10 +18,10 @@ package io.micronaut.http.server.netty.jackson;
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.http.MediaType;
 import io.micronaut.http.annotation.Consumes;
-import io.micronaut.http.server.HttpServerConfiguration;
 import io.micronaut.http.server.netty.HttpContentProcessor;
 import io.micronaut.http.server.netty.HttpContentSubscriberFactory;
 import io.micronaut.http.server.netty.NettyHttpRequest;
+import io.micronaut.http.server.netty.configuration.NettyHttpServerConfiguration;
 import io.micronaut.json.JsonMapper;
 import jakarta.inject.Singleton;
 
@@ -36,7 +36,7 @@ import jakarta.inject.Singleton;
 @Internal
 public class JsonHttpContentSubscriberFactory implements HttpContentSubscriberFactory {
 
-    private final HttpServerConfiguration httpServerConfiguration;
+    private final NettyHttpServerConfiguration httpServerConfiguration;
     private final JsonMapper jsonMapper;
 
     /**
@@ -45,7 +45,7 @@ public class JsonHttpContentSubscriberFactory implements HttpContentSubscriberFa
      */
     public JsonHttpContentSubscriberFactory(
             JsonMapper jsonMapper,
-            HttpServerConfiguration httpServerConfiguration) {
+            NettyHttpServerConfiguration httpServerConfiguration) {
         this.httpServerConfiguration = httpServerConfiguration;
         this.jsonMapper = jsonMapper;
     }
